Add firebase gitignored config
