+++ lit/src/shared/firebase_config_example.ts
@@ -1,0 +1,10 @@
+import {FirebaseOptions} from 'firebase/app';
+
+export const FIREBASE_CONFIG: FirebaseOptions = {
+  apiKey: 'your-api-key',
+  authDomain: 'your-auth-domain',
+  projectId: 'your-project-id',
+  storageBucket: 'your-storage-bucket',
+  messagingSenderId: 'your-messaging-sender-id',
+  appId: 'your-app-id',
+};
